requirements.txt
django==4.2.0
gunicorn==21.2.0
python-dotenv==1.0.1
waitress==2.1.2
whitenoise==6.6.0